net: l2switch: Do not send "no carrier" notification to eth0 for L2 switch
The old NXP driver ported from 2.6.x tree has an issue when one wants
to setup transmission via switch to one board (by using mac0) and removes
cable for mac1.
Without this patch the eth0 is notified about losing carrier and the
communication on mac0 also is stopped.
As this driver is not using (yet) DSA infrastructure - do not send the
carrier info for L2 switch eth interface.

The only drawback is that eth0 behaves as it would be a "fixed-link" phy
with being UP (ip link show eth0) even when no cable is attached.

Signed-off-by: Lukasz Majewski <lukma@denx.de>