net: mtip: Use proper network devices when offloading is enabled
Before this patch, when bridge offloading was enabled (ie. fep->br_offload
was 1), the received packet was "handled" by the network device assigned
to NAPI.

This was problematic, when NAPI was initalized with lan0 assigned network
device, but it was not operational (i.e. the case when board boots with
lan0 cable not connected and only lan1 is plugged in).

With current code - packet is handled by lan0/lan1, which are bridged.
In that way packets are not lost.

Last, but not least - as "switching" is done in HW, packets are send
through switch without issues.

Signed-off-by: Lukasz Majewski <lukma@denx.de>