net: ptp: Adjust fec_ptp.c to support its operation FEC and MTIP drivers
The fec_ptp.c is a common file used by both FEC and MTIP L2 SW drivers.
The trick here is to introduce typedev fep_t, which will be different
depending on the target device.

It was also necessary to adjust the way of getting the *fep pointer itself
as on L2 switch driver it is accessible from both ports' network device
structure private member.

Signed-off-by: Lukasz Majewski <lukma@denx.de>