net: mtip: vf: Enable Descriptor Byte Swap for vf610 SoC
The Vybrid vf610 ENET has special bit - i.e. DBSWP (bit 8) in ENETx_ECR
register. It is responsible for swapping buffer descriptor bytes to
support little-endian devices.

This bit is not present on imx287 ENET (part of area of reserved bits),
but writing to it doesn't harm (as it is done in fec_main.c).

Signed-off-by: Lukasz Majewski <lukasz.majewski@mailbox.org>