net: mtip: Set offload_fwd_mark when L2 offloading is enabled
The packet's skb->offload_fwd_mark needs to be set according to the
state of br_offload flag.

It is required to mark ingress packets as already in-HW offloaded to
avoid forwarding (or flooding) them to "other" bridge ports by the
bridge driver itself.

This approach avoids problems with packets duplication (especially
broadcast ones).

Signed-off-by: Lukasz Majewski <lukasz.majewski@mailbox.org>